refactor(frontend): split organization-settings into smaller components
Extract OrganizationTeams, OrganizationMembers, and
OrganizationInvitations into dedicated components to reduce the
size of the organization settings page.
